+++ src/auth/pages/RegisterPage.jsx
@@ -64,7 +64,10 @@ export const RegisterPage = () => {
 
   return (
     <AuthLayout title="Crear una cuenta">
-      <form onSubmit={handleSubmit}>
+      <form
+        onSubmit={handleSubmit}
+        className="animate__animated animate__fadeIn animate__faster"
+      >
         <Grid container>
           <Grid item xs={12} sx={{ mt: 2 }}>
             <TextField
