+++ src/firebase/providers.js
@@ -94,3 +94,7 @@ export const loginWithEmailAndPassword = async ({ email, password }) => {
     };
   }
 };
+
+export const logoutFirebase = async () => {
+  return await firebaseAuth.signOut();
+};
