+++ src/journal/components/NavBar.jsx
@@ -1,8 +1,16 @@
 import PropTypes from "prop-types";
 import { LogoutOutlined, MenuOutlined } from "@mui/icons-material";
 import { AppBar, Grid, IconButton, Toolbar, Typography } from "@mui/material";
+import { useDispatch } from "react-redux";
+import { startFirebaseLogout } from "../../store/auth/thunks";
 
 export const NavBar = ({ drawerWidth }) => {
+  const dispatch = useDispatch();
+
+  const handleLogout = () => {
+    dispatch(startFirebaseLogout());
+  };
+
   return (
     <AppBar
       position="fixed"
@@ -29,7 +37,7 @@ export const NavBar = ({ drawerWidth }) => {
           <Typography variant="h6" noWrap component="div">
             JournalApp
           </Typography>
-          <IconButton color="error">
+          <IconButton color="error" onClick={handleLogout}>
             <LogoutOutlined />
           </IconButton>
         </Grid>
