+++ src/journal/components/SideBar.jsx
@@ -13,8 +13,11 @@ import {
   Typography,
 } from "@mui/material";
 import { TurnedInNot } from "@mui/icons-material";
+import { useSelector } from "react-redux";
 
 export const SideBar = ({ drawerWidth }) => {
+  const { displayName } = useSelector((state) => state.auth);
+
   return (
     <Box
       component="nav"
@@ -33,7 +36,7 @@ export const SideBar = ({ drawerWidth }) => {
       >
         <Toolbar>
           <Typography variant="h6" noWrap component="div">
-            Florencia Luz Duarte
+            {displayName}
           </Typography>
           <Divider />
         </Toolbar>
