+++ src/journal/views/NoteView.jsx
@@ -5,6 +5,7 @@ import { ImageGallery } from "../components";
 export const NoteView = () => {
   return (
     <Grid
+      className="animate__animated animate__fadeIn animate__faster"
       container
       direction="row"
       justifyContent="space-between"
