+++ src/journal/views/NothingSelectedView.jsx
@@ -4,6 +4,7 @@ import { StarOutline } from "@mui/icons-material";
 export const NothingSelectedView = () => {
   return (
     <Grid
+      className="animate__animated animate__fadeIn animate__faster"
       container
       spacing={0}
       direction="column"
