+++ src/router/AppRouter.jsx
@@ -1,24 +1,11 @@
 import { Navigate, Route, Routes } from "react-router-dom";
 import { AuthRoutes } from "../auth/routes/AuthRoutes";
-import { firebaseAuth } from "../firebase/config";
 import { JournalRoutes } from "../journal/routes/JournalRoutes";
-import { useDispatch, useSelector } from "react-redux";
 import { CheckingAuthentication } from "../ui";
-import { useEffect } from "react";
-import { onAuthStateChanged } from "firebase/auth";
-import { login, logout } from "../store/auth";
+import { useCheckAuth } from "../hooks/useCheckAuth";
 
 export const AppRouter = () => {
-  const { status } = useSelector((store) => store.auth);
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    onAuthStateChanged(firebaseAuth, async (user) => {
-      if (!user) return dispatch(logout());
-      const { uid, displayName, email, photoURL } = user;
-      dispatch(login({ uid, displayName, email, photoURL }));
-    });
-  }, [dispatch]);
+  const { status } = useCheckAuth();
 
   if (status === "checking") {
     return <CheckingAuthentication />;
